Feat: "댓글 기능,어드민 권한 설정"

+++ springblog-main/src/main/java/com/sparta/board/service/BoardService.java
@@ -1,14 +1,16 @@
 package com.sparta.board.service;
 
-import com.sparta.board.dto.DeleteDto;
+import com.sparta.board.entity.User;
 import com.sparta.board.dto.MessageResponseDto;
 import com.sparta.board.dto.PostRequestDto;
 import com.sparta.board.dto.PostResponseDto;
 import com.sparta.board.entity.Board;
 import com.sparta.board.entity.Cmt;
+import com.sparta.board.entity.UserRoleEnum;
 import com.sparta.board.jwt.JwtUtil;
 import com.sparta.board.repository.BoardRepository;
 import com.sparta.board.repository.CmtRepository;
+import com.sparta.board.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
@@ -17,48 +19,75 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class BoardService {
-    private final CmtRepository cmtRepository;
+
     private final BoardRepository boardRepository;
     private final JwtUtil jwtUtil;
+    private final CmtRepository cmtRepository;
+    private final UserRepository userRepository;
+
+
     @Transactional
-    public PostResponseDto createPost(PostRequestDto requestDto) {
-//        String username=jwtUtil.getUsername();
-        String username=requestDto.getUsername();
-//        if(!username.equals(requestDto.getUsername())){
-//            throw new IllegalArgumentException("사용자명이 일치하지 않습니다.");
-//        }
-        Board board = new Board(requestDto);
+    public PostResponseDto createPost(PostRequestDto requestDto, HttpServletRequest req) {
+        String username= jwtUtil.getUsername();
+        Board board = new Board(requestDto,username);
         Board savePost = boardRepository.save(board);
         return new PostResponseDto(savePost);
     }
+
+    @Transactional
      public List<PostResponseDto> getPosts() {
+        for(Long i=1L;i<= boardRepository.count();i++){
+            Board board = findPost(i);
+            List<Cmt> updatedCommentList = getcmt(i);
+            board.update(updatedCommentList);
+        }
         return boardRepository.findAllByOrderByModifiedAtDesc()
                 .stream()
                 .map(PostResponseDto::new)
                 .collect(Collectors.toList());
     }
-     @Transactional
-     public PostResponseDto getPost(Long id) {
+    public List<Cmt> getcmt(Long id) {
+        return cmtRepository.findAllBypostid(id);
+    }
+    @Transactional
+    public PostResponseDto getPost(Long id) {
         Board board = findPost(id);
+        List<Cmt> updatedCommentList = getcmt(id);
+        board.update(updatedCommentList);
         return new PostResponseDto(board);
     }
 
     @Transactional
-    public PostResponseDto updatePost(Long id, PostRequestDto requestDto, HttpServletRequest req) {
+    public MessageResponseDto updatePost(Long id, PostRequestDto requestDto, HttpServletRequest req) {
         Board board = findPost(id);
-        board.update(requestDto);
-        return new PostResponseDto(board);
+        Optional <User> user=userRepository.findByUsername(jwtUtil.getUsername());
+        UserRoleEnum role=user.get().getRole();
+        if((jwtUtil.getUsername().equals(board.getUsername()))||role.equals(UserRoleEnum.ADMIN) ){
+            board.update(requestDto);
+            return new MessageResponseDto("수정 완료", HttpStatus.OK.value());
+        }
+        else{
+            return new MessageResponseDto("수정 실패", HttpStatus.EXPECTATION_FAILED.value());
+        }
     }
 
-    public MessageResponseDto deletePost(Long id, PostRequestDto requestDto, HttpServletRequest req) {
+    public MessageResponseDto deletePost(Long id) {
         Board board = findPost(id);
+        Optional <User> user=userRepository.findByUsername(jwtUtil.getUsername());
+        UserRoleEnum role=user.get().getRole();
+        if((jwtUtil.getUsername().equals(board.getUsername()))||role.equals(UserRoleEnum.ADMIN) ){
         boardRepository.delete(board);
         return new MessageResponseDto("삭제 완료", HttpStatus.OK.value());
+        }
+        else{
+            return new MessageResponseDto("삭제 실패", HttpStatus.EXPECTATION_FAILED.value());
+        }
     }
 
     private Board findPost(Long id){
@@ -66,20 +95,5 @@ public class BoardService {
                 new IllegalArgumentException("선택한 게시글은 존재하지 않습니다."));
     }
 
-//    private void validateTokenAndUser(PostRequestDto requestDto, HttpServletRequest req){
-//        String token = jwtUtil.getJwtFromHeader(req);
-//        if(!jwtUtil.validateToken(token)){
-//            throw new IllegalArgumentException("유효하지 않은 토큰입니다.");
-//        }
-//        Claims claims = jwtUtil.getUserInfoFromToken(token);
-//        String username = claims.getSubject();
-//        if(!username.equals(requestDto.getUsername())){
-//            throw new IllegalArgumentException("사용자명이 일치하지 않습니다.");
-//        }
-//    }
-
-    public List<Cmt> getcmt(Long id) {
-        return cmtRepository.findAllBypostid(id);
-    }
 
 }
