+++ springblog-main/src/main/java/com/sparta/board/service/CmtService.java
@@ -2,9 +2,11 @@ package com.sparta.board.service;
 
 import com.sparta.board.dto.CmtRequestDto;
 import com.sparta.board.dto.CmtResponseDto;
+import com.sparta.board.dto.MessageResponseDto;
 import com.sparta.board.dto.PostResponseDto;
 import com.sparta.board.entity.Board;
 import com.sparta.board.entity.Cmt;
+import com.sparta.board.filter.AuthFilter;
 import com.sparta.board.jwt.JwtUtil;
 import com.sparta.board.repository.BoardRepository;
 import com.sparta.board.repository.CmtRepository;
@@ -12,6 +14,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,66 +26,47 @@ public class CmtService {
     private final CmtRepository cmtRepository;
     private final JwtUtil jwtUtil;
     private final BoardRepository boardRepository;
-//     public CmtResponseDto createcmt(CmtRequestDto cmtRequestDto) {
-//
-//        String username="";
-//        Long id= cmtRequestDto.getPostid();
-//        Cmt cmt= new Cmt(cmtRequestDto,username);
-//        Cmt savecmt=cmtRepository.save(cmt);
-//        Board board =findPost(id);
-//        List<Cmt> updatedCommentList = getcmt(id);
-//        board.setList(updatedCommentList);
-//        return new CmtResponseDto(savecmt);
-//    }
-//    @Transactional
-//    public CmtResponseDto createcmt(CmtRequestDto cmtRequestDto) {
-//        String username = "";
-//        Long id = cmtRequestDto.getPostid();
-//        Cmt cmt = new Cmt(cmtRequestDto, username);
-//        Cmt savecmt = cmtRepository.save(cmt);
-//        Board board = findPost(id);
-//        List<Cmt> updatedCommentList = getcmt(id);
-//        updatedCommentList.add(cmt); // 새로운 Cmt 객체를 commentlist에 추가
-//        board.setList(updatedCommentList);
-//
-//        boardRepository.save(board); // board 객체 저장하여 변경 내용을 영속성 컨텍스트에 반영
-//
-//        return new CmtResponseDto(savecmt);
-//
-//    }
-
+    private final AuthFilter authFilter;
     @Transactional
     public CmtResponseDto createcmt(CmtRequestDto cmtRequestDto) {
-        String username = "";
+
+        String username = jwtUtil.getUsername();
         Long id = cmtRequestDto.getPostid();
          Cmt cmt = new Cmt(cmtRequestDto, username);
         Cmt savecmt = cmtRepository.save(cmt);
         return new CmtResponseDto(savecmt);
     }
 
-
-    public PostResponseDto insertcmt(Long id){
-    Board board = findPost(id);
-    List<Cmt> updatedCommentList = getcmt(id);
-//    board.setList(updatedCommentList);
-//    boardRepository.save(board);
-    board.update(updatedCommentList);
-    PostResponseDto responseDto =new PostResponseDto(board);
-    return responseDto;
-    }
-
+//    public PostResponseDto insertcmt(Long id){
+//        Board board = findPost(id);
+//        List<Cmt> updatedCommentList = getcmt(id);
+//        board.update(updatedCommentList);
+//        boardRepository.save(board);
+//        PostResponseDto responseDto =new PostResponseDto(board);
+//        return responseDto;
+//    }
     public List<Cmt> getcmt(Long id) {
         return cmtRepository.findAllBypostid(id);
     }
-
-     public CmtResponseDto updatecmt(Long id, CmtRequestDto cmtRequestDto) {
+     public MessageResponseDto updatecmt(Long id, CmtRequestDto cmtRequestDto) {
         Cmt cmt = findcmt(id);
-        cmt.update(cmtRequestDto);
-        return new CmtResponseDto(cmt);
+        if(cmt.getUsername().equals(jwtUtil.getUsername())) {
+            cmt.update(cmtRequestDto);
+            return new MessageResponseDto("수정 완료", HttpStatus.OK.value());
+        }
+        else{
+            return new MessageResponseDto("수정 불가", HttpStatus.EXPECTATION_FAILED.value());
+        }
     }
-     public void deletecmt(Long id) {
+     public MessageResponseDto deletecmt(Long id) {
         Cmt cmt =findcmt(id);
-        cmtRepository.delete(cmt);
+         if(cmt.getUsername().equals(jwtUtil.getUsername())) {
+             cmtRepository.delete(cmt);
+             return new MessageResponseDto("삭제 완료", HttpStatus.OK.value());
+         }
+         else{
+             return new MessageResponseDto("삭제 불가", HttpStatus.EXPECTATION_FAILED.value());
+         }
     }
 
     private Cmt findcmt(Long id){
